More refactoring of utilities
