+++ apps/web/app/globals.css
@@ -125,10 +125,6 @@
   }
 }
 
-/*
-  ---break---
-*/
-
 @layer base {
   * {
     @apply border-border outline-ring/50;
@@ -136,49 +132,6 @@
   body {
     @apply bg-background text-foreground;
   }
-}
-
-/* Accessibility improvements */
-
-/* Screen reader only utility - hidden visually but available to screen readers */
-.sr-only {
-  position: absolute;
-  width: 1px;
-  height: 1px;
-  padding: 0;
-  margin: -1px;
-  overflow: hidden;
-  clip: rect(0, 0, 0, 0);
-  white-space: nowrap;
-  border-width: 0;
-}
-
-.sr-only:focus {
-  position: static;
-  width: auto;
-  height: auto;
-  padding: inherit;
-  margin: inherit;
-  overflow: visible;
-  clip: auto;
-  white-space: normal;
-}
-
-/* Focus styles for better keyboard navigation */
-:focus {
-  outline: 2px solid #3b82f6;
-  outline-offset: 2px;
-}
-
-:focus-visible {
-  outline: 2px solid #3b82f6;
-  outline-offset: 2px;
-}
-
-/* High contrast focus for dark mode */
-.dark :focus,
-.dark :focus-visible {
-  outline-color: #60a5fa;
 }
 
 /* Skip link focus styles */
