+++ apps/web/components/ui/dialog.tsx
@@ -4,7 +4,7 @@ import * as React from "react";
 import * as DialogPrimitive from "@radix-ui/react-dialog";
 import { XIcon } from "lucide-react";
 
-import { cn } from "@/utils";
+import { cn } from "@/utils/ui";
 
 function Dialog({
   ...props
