+++ apps/web/components/ui/input.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 
-import { cn } from "@/utils";
+import { cn } from "@/utils/ui";
 
 export type InputProps = React.InputHTMLAttributes<HTMLInputElement>;
