+++ apps/web/components/ui/skeleton.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/utils";
+import { cn } from "@/utils/ui";
 
 function Skeleton({
   className,
