--- apps/web/utils/constants/index.ts
@@ -1,2 +1,0 @@
-// Re-export all constants for easier imports
-export * from './app';
