+++ apps/web/utils/fetch-stats-data.ts
@@ -1,5 +1,9 @@
 import { createClientForServer } from "@/utils/supabase/server";
-import { ITEM_TYPE_ORDER, CHART_CONFIG, ITEM_TYPES } from "@/utils/constants/app";
+import {
+  ITEM_TYPE_ORDER,
+  CHART_CONFIG,
+  ITEM_TYPES,
+} from "@/utils/constants/app";
 import { getCurrentYear } from "@/utils/formatters/date";
 import { ItemCountEntry } from "@/types";
 
@@ -31,13 +35,19 @@ export async function fetchStatsData() {
   ) =>
     itemCountsData
       ?.filter((entry): entry is ItemCountEntry =>
-        Object.values(ITEM_TYPES).includes(entry.itemtype as any),
+        Object.values(ITEM_TYPES).includes(entry.itemtype),
       )
-      .sort((a, b) => ITEM_TYPE_ORDER[a.itemtype as keyof typeof ITEM_TYPE_ORDER] - ITEM_TYPE_ORDER[b.itemtype as keyof typeof ITEM_TYPE_ORDER])
+      .sort(
+        (a, b) =>
+          ITEM_TYPE_ORDER[a.itemtype as keyof typeof ITEM_TYPE_ORDER] -
+          ITEM_TYPE_ORDER[b.itemtype as keyof typeof ITEM_TYPE_ORDER],
+      )
       .map((entry) => ({
         type: entry.itemtype,
         count: entry[key],
-        fill: CHART_CONFIG[entry.itemtype as keyof typeof CHART_CONFIG]?.color || "hsl(var(--chart-1))",
+        fill:
+          CHART_CONFIG[entry.itemtype as keyof typeof CHART_CONFIG]?.color ||
+          "hsl(var(--chart-1))",
       }));
 
   return {
