--- apps/web/utils/formatters/index.ts
@@ -1,2 +1,0 @@
-// Re-export all formatters for easier imports
-export * from './date';
