+++ apps/web/utils/server/redirects.ts
@@ -1,6 +1,4 @@
 import { redirect } from "next/navigation";
-import { clsx, type ClassValue } from "clsx";
-import { twMerge } from "tailwind-merge";
 
 /**
  * Redirects to a specified path with an encoded message as a query parameter.
@@ -15,8 +13,4 @@ export function encodedRedirect(
   message: string,
 ): never {
   return redirect(`${path}?${type}=${encodeURIComponent(message)}`);
-}
-
-export function cn(...inputs: ClassValue[]) {
-  return twMerge(clsx(inputs));
 }
