+++ apps/web/utils/validators/index.ts
@@ -1,2 +1,2 @@
 // Re-export all validators for easier imports
-export * from './item';
+export * from "./item";
